+++ CHANGELOG.rst
@@ -10,6 +10,7 @@ This project adheres to `Semantic Versioning <http://semver.org/>`_.
 - Added unit and range verification checks to `ApplyPointingModel`command
 - Added in Read/Write dscPowerLimitKw attribute to be used when SetStandbyFPMode, SetStandbyLPMode and Slew are invoked.
 - Upgraded ska-mid-dish-ds-manager chart to v2.2.0
+- Updated SPFRx operatingMode from DATA_CAPTURE to OPERATE to match Rev 4 ICD
 - Added `Abort` tango command which cancels any task and restores the dish to FP mode
 
   - `AbortCommmands` implements the same handler as `Abort`